Start reworking game code
I put compile-time flags around the capture game, the QSO game, and mastermind.

The code builds with each of those flags on and off, but the results have not been tested on a badge.

The conditional compilation changed the data format of the state file, so state can't persist correctly across different build options.

There's no capture game code to speak of yet, this just adds some defines for gameplay and state get/set