remove the check when starting a notification
The check during capture_process_heard to make sure it's ok to display a notification
could be invalid because of a race condition. It really only matters whether it's ok to
display a notification during the callback where we display it, and the callback is only
supposed to be called from places where that's safe.