Don't call app_sched_execute() from the capture timer handler
The code was periodically crashing or restarting after notifications were displayed.
In at least one case, this was seen to be related to the call of app_sched_execute() from
the capture timer handler, which immediately called the state save timer handler, which
uses a lot of stack (because it holds the entire state structure).

There was also some bad coding that made it possible for the selection of a creature id
to take a long time, which could have th eeffect of causing the timer queue to back up,
since this is called from the capture game timer handler. That was made better.