Rearrange creatre data reads into a single thread
Crashes and file read failures were happening because fatfs access to the same file can't be re-entrant.
Moved the file reads from the notification callback into the capture timer handler, and added a new state
to notifications to allow that processing.

Also, keep the last creature in the notification state when finished, and use it to prevent
notifying for the same creatre twice in a row. This was happening if the user acknowledged a notification
while the creature advertisements were still being sent.